file_transformation_service/requirements.txt
et-xmlfile==1.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
PyPDF2==2.11.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
typing_extensions==4.4.0
